Enable EdmReference, EdmInclude vocabulary annotation
