+++ src/Microsoft.OData.Edm/Csdl/CsdlJsonWriter.cs
@@ -77,107 +77,10 @@ namespace Microsoft.OData.Edm.Csdl
         /// </summary>
         private void WriteReferenceElements()
         {
-            // The document object MAY contain the member $Reference to reference other CSDL documents.
-            IEnumerable<IEdmReference> references = model.GetEdmReferences();
-            if (references != null && references.Any())
-            {
-                // The value of $Reference is an object that contains one member per referenced CSDL document.
-                this.jsonWriter.WritePropertyName("$Reference");
-                this.jsonWriter.WriteStartObject();
+            EdmModelReferenceElementsJsonVisitor visitor
+                = new EdmModelReferenceElementsJsonVisitor(this.model, this.jsonWriter, this.edmxVersion);
 
-                // The value of $Reference is an object that contains one member per referenced CSDL document.
-                foreach (IEdmReference reference in references)
-                {
-                    // The name of the pair is a URI for the referenced document.
-                    this.jsonWriter.WritePropertyName(reference.Uri.OriginalString);
-
-                    // The value of each member is a reference object.
-                    this.jsonWriter.WriteStartObject();
-
-                    WriteIncludes(reference.Includes);
-
-                    WriteIncludeAnnotations(reference.IncludeAnnotations);
-
-                    // The reference object MAY contain annotations, so far it's not supported.
-                    // End of each reference document.
-                    this.jsonWriter.WriteEndObject();
-                }
-
-                // End of $Reference
-                this.jsonWriter.WriteEndObject();
-            }
-        }
-
-        /// <summary>
-        /// $Include Array.
-        /// </summary>
-        private void WriteIncludes(IEnumerable<IEdmInclude> includes)
-        {
-            if (includes == null || !includes.Any())
-            {
-                return;
-            }
-
-            // The reference object MAY contain the members $Include
-            this.jsonWriter.WritePropertyName("$Include");
-
-            // The value of $Include is an array.
-            this.jsonWriter.WriteStartArray();
-
-            foreach (IEdmInclude include in includes)
-            {
-                // Array items are objects.
-                this.jsonWriter.WriteStartObject();
-
-                // MUST contain the member $Namespace
-                this.jsonWriter.WriteRequiredProperty("$Namespace", include.Namespace);
-
-                // MAY contain the member $Alias
-                this.jsonWriter.WriteOptionalProperty("$Alias", include.Alias);
-
-                // The include object MAY contain annotations.
-                // So far, it's not supported in ODL.
-
-                this.jsonWriter.WriteEndObject();
-            }
-
-            this.jsonWriter.WriteEndArray();
-        }
-
-        /// <summary>
-        /// $IncludeAnnotations Array.
-        /// </summary>
-        private void WriteIncludeAnnotations(IEnumerable<IEdmIncludeAnnotations> includeAnnotations)
-        {
-            if (includeAnnotations == null || !includeAnnotations.Any())
-            {
-                return;
-            }
-
-            // The reference object MAY contain the members $IncludeAnnotations
-            this.jsonWriter.WritePropertyName("$IncludeAnnotations");
-
-            // The value of $IncludeAnnotations is an array.
-            this.jsonWriter.WriteStartArray();
-
-            foreach (IEdmIncludeAnnotations includeAnnotation in includeAnnotations)
-            {
-                // Array items are objects
-                this.jsonWriter.WriteStartObject();
-
-                // MUST contain the member $TermNamespace
-                this.jsonWriter.WriteRequiredProperty("$TermNamespace", includeAnnotation.TermNamespace);
-
-                // MAY contain the members $Qualifier
-                this.jsonWriter.WriteOptionalProperty("$Qualifier", includeAnnotation.Qualifier);
-
-                // MAY contain the members $TargetNamespace
-                this.jsonWriter.WriteOptionalProperty("$TargetNamespace", includeAnnotation.TargetNamespace);
-
-                this.jsonWriter.WriteEndObject();
-            }
-
-            this.jsonWriter.WriteEndArray();
+            visitor.VisitEdmReferences(this.model);
         }
 
         /// <summary>
